refactor: modularity and cleaner implementation with usRef
